third_party/blink/web_tests/fast/forms/state-restore-skip-stateless-expected.txt

CONSOLE MESSAGE: Test if state of stateless form control types are not saved.
PASS There is a state for text
PASS There is a state for tel
PASS There is a state for url
PASS There is a state for email
PASS There is a state for number
PASS There is a state for range
PASS There is a state for checkbox
PASS There is a state for radio
PASS There is a state for file
PASS There is a state for select-one
PASS There is a state for select-multiple
PASS There is a state for textarea
PASS There is no state for fieldset
PASS There is no state for hidden
PASS There is no state for password
PASS There is no state for submit
PASS There is no state for reset
PASS There is no state for button
PASS There is no state for output
PASS successfullyParsed is true

TEST COMPLETE
